Code for mixing the input for the tracks
The speed of the tank tracks can be calculated with the use of this code. This file isn't supposed to be included, nor is it supposed to work on its own. It's just a temporary file in which the code is stored waiting to be implemented into the right class.