Backend Overhaul + bugfixes
The listener was not only listening, but also parsing and controlling the robot depending on said parsed data. The result of this chaos was that the listener was acting as the controller, which meant that the vision wouldn't have a proper way of connecting to the arm and tank tracks.
- arm hasn't been fully tested yet, but seems to work as it's supposed to.
- talker hasn't been tested at all, but it didn't undergo (mayor) changes either.
- vision hasn't been tested yet, but didn't undergo any changes.
- listener has been tested, and looks like it's working fine (though, it seems to be receiving way more than before. This could explain why we didn't receive all events)
- tanktracks have been fully tested and work as supposed. (deadzone should be changed to 0.2 though)