Arm Freeze Fix
Arm was freezing. This was most likely caused by mutex lock issues. Delays and locks have been tweaked to fix this.
Another fix is for the stop button. Occasionally the parsed_input will be updated twice while the controller is still busy working on the older data causing the signal to be 'ignored'.
FOUND BUG: while pressing 'refresh' on the controller it will cause a segmentation fault. (possible causes: while reading value || while sending data to the controller)